requirements-full.txt
# Full requirements including Google Drive integration
# Use this if you need Google Drive functionality

# Core dependencies for FCR Dashboard
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
openpyxl>=3.1.0

# Google Drive integration
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

